mate-dict.pc: depend upon proper gtk version
If the package is built against gtk-3.0 (as in
./configure --with-gtk=3.0), mate-dict.pc still wants gtk-2.0
since it's hardcoded it there. This can lead to improper builds
of packages using libmatedict (although I can't find any now),
as well as improper RPM dependencies, like this:

	$ rpm -e gtk2-devel
	error: Failed dependencies:
	...
	pkgconfig(gtk+-2.0) is needed by (installed) mate-utils-devel-1.14.0-1gtk3.fc23.x86_64

Fortunately, the file is already generated, so it's trivial to add
a proper substitution.

[v2: also add min GTK version required]

Signed-off-by: Kir Kolyshkin <kir@openvz.org>