Dictionary: Fix Gtk3.21 segfault
Fix GTK 3.21 segfault caused by issue with calling gtk_widget_realize at this point in the program. Show the widget instead as this does the same thing implicitly and doesn't segfault. Restrict this to GTK 3.21 for now. Mate-dictionary windows haven't resized to fit text all the way back to MATE 1.10 as tested with GTK 3.14, so this whole function needs to either be fixed or removed. My earlier prevented it from being called by disabling a function that used GtkStyle but in this case GtkStyle isn't causing trouble -yet.