Gtk 3.21 mate-search-tool: fix short entry box
Recent changes in GTK 3.21 (probably  https://git.gnome.org/browse/gtk+/commit/?id=a72f1c76c87de7a8124a809fe9692194273c563e ) have caused the "Name Contains" entry and "Look in Folder" button to render far shorter than normal. GNOME has changed the way Gtktable (deprecated) computes expansion  and the table itself does not want to expand to fill the box.

For some reason this does not happen to gnome-search-tool but I was unable to find the difference responsible as the code is very close. Gnome-search-tool still uses Gtktable too as they have not updated this file in something like 5 years.   For mate-search-tool, adding a style class and a cssprovider forces the combobox containing the entry and button back to normal size, pushing the table back out as well. When I tried it on the table in initial tests in my theme it did not work so I put it on the combobox where it did.